# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Universal Web-to-Markdown & B2B Lead Intelligence API",
    "description": "Production-ready edge API for AI Agents, LLM RAG pipelines, and B2B Automation. Converts websites into clean, token-efficient Markdown while extracting business leads, metadata, and links in real-time.",
    "version": "1.1.0",
    "contact": {
      "email": "top.ai.saas@gmail.com"
    }
  },
  "servers": [
    {
      "url": "https://web-to-markdown-api.topaisaas.workers.dev",
      "description": "Cloudflare Global Edge Server"
    }
  ],
  "paths": {
    "/v1/extract": {
      "post": {
        "summary": "Full Extraction: Clean Markdown + B2B Leads + Metadata",
        "description": "Extracts clean, noise-free Markdown optimized for LLM token savings alongside business contacts (emails, phone numbers, socials).",
        "operationId": "extractFull",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "example": "https://stripe.com"
                  },
                  "extract_leads": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful full extraction"
          }
        }
      }
    },
    "/v1/extract/leads": {
      "post": {
        "summary": "Lead Hunter: Extract Emails, Phones & Social Profiles",
        "description": "Dedicated high-speed endpoint for sales and lead generation agents. Returns all discovered business contacts without processing markdown.",
        "operationId": "extractLeadsOnly",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "example": "https://stripe.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful lead extraction"
          }
        }
      }
    },
    "/v1/extract/metadata": {
      "post": {
        "summary": "SEO & Social Metadata: Title, Description, OpenGraph, Favicon",
        "description": "Instant (<10ms) extraction of page metadata, OpenGraph tags, language, and favicons for link preview widgets and SEO crawlers.",
        "operationId": "extractMetadataOnly",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "example": "https://stripe.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful metadata extraction"
          }
        }
      }
    },
    "/v1/extract/links": {
      "post": {
        "summary": "Crawler & Site Explorer: Internal, External & Document Links",
        "description": "Discovers all hyperlinks on a target page, cleanly categorized into internal site links, external references, and downloadable documents (PDF, DOCX, XLSX).",
        "operationId": "extractLinksOnly",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "example": "https://stripe.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful link extraction"
          }
        }
      }
    },
    "/v1/health": {
      "get": {
        "summary": "Service Healthcheck",
        "responses": {
          "200": {
            "description": "Service is healthy"
          }
        }
      }
    }
  }
}